fix(backend/graph): move toUserModel to helpers.go so gqlgen regen keeps it
